Added an option to delete files before the upload
In some cases it may be important to make sure the files we upload won't reuse the same inode, i.e. in case the file(s) with the same name already exists in the target directory and have hard links.

Currently uploading a new file with the same name will silently push a change to all linked instances, something we may want to avoid.

Deleting the files before the upload takes care of this.

Signed-off-by: Lev Veyde <lveyde@redhat.com>