Implemented source side wildcard (globbing) processing
Previously, if the source included any wildcards, these were processed at the target side, thus
potentially deleting more files that were later uploaded, as the same wildcard selection could
be expanded to a different set of files.

This patch modified the code to perform the expansion at the source side, thus only deleting
the same list of files that are later will be uploaded.

It also updates the test and README.md file for the change.

Signed-off-by: Lev Veyde <lveyde@redhat.com>